Data structure re-design
We have two lists fragments fragment_by_start, fragment_by_end
as a handle to dynamically access the underlying memory

In javascrript deleting an entry, from one list automatically
invalidated the entry in the other list.

I am not happy with the fix here, but the obvious solution
is to cross-link and then delete entries in fragment_by_end and
fragment_by_start as a pair.

At the moment I am aiming to get the test passing, then from a
position of stability I may have to revisit this.

This is an initial, buggy version of the cross-link pattern.