Removed package dependency "derivative"
I like the module, it is repeatly used throught the code-base

But it has not been updated in two years.

Running Cargo outdated in that package give a very long list
A package using a lot of outdated packages is unsecure.

Also it triggers warning about syn, in my crates because of the
derivative package both syn1.0 and syn2.0 need to been included
leading to bloat in my packages.